refactor: remove method from write script
